Minor fixes un sort button (Subcategory) and options buttons (Advertisement)
Changes:

1. Now the menu with the possible options to sort, is correctly
displayed behind the “Sort by” button. Additionally, now if the user
places the cursor above, the cursor is a pointer.

In order to do so, the “ul” and “li” tags has been replaced by “div”,
otherwise, there was always a small margin without any option to remove
it.

2. A slightly grey has been added as background color in the
Advertisement page.

3. Now all footers has EXACTLY the same text.